refactor: move SupportedModelTypes under datamodels
