+++ viewer/src/datamodels/base/SupportedModelTypes.ts
@@ -1,0 +1,4 @@
+/*!
+ * Copyright 2020 Cognite AS
+ */
+export type SupportedModelTypes = 'pointcloud' | 'cad';
